# services/mapi-xbt/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bitcoin - Blockchain Indexer API",
    "description": "This API provides core indexer endpoints with support for Bitcoin metaprotocols by delivering real-time, rollback-protected access to Bitcoin's UTXO data, enabling developers to build responsive and reliable blockchain applications without managing complex infrastructure.\n\n#### Key Features:\n- **Real-Time Data with Rollback Protection:** Ensures data accuracy by handling chain reorganizations gracefully, providing live data without sacrificing integrity.\n- **Comprehensive UTXO Indexing:** Specialized pipelines extract, match, and process on-chain information, including handling rollbacks, to provide accurate and up-to-date data.\n\n#### Key Benefits for Developers:\nBy abstracting the complexities of blockchain data retrieval and processing, the Bitcoin Indexer API empowers developers to focus on building innovative applications with confidence in fast and reliable access to historical chain data.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "v0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/addresses/{address}/inscriptions/activity": {
      "get": {
        "tags": [
          "Addresses"
        ],
        "summary": "Inscription Activity by Address",
        "description": "Returns all inscription-related transactions involving a specific address. Can be filtered by activity type (send, receive, self-transfer), narrowed to a specific inscription, and sorted chronologically. Useful for building dashboards, tracking user behavior, or filtering unwanted spam activity.",
        "operationId": "inscription_activity_by_address",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Bitcoin address or hex encoded script pubkey",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "bc1p27j3fa2mr3d50m3uaavr0ntyzr0v2a27n48lc9gxpkzd4xye6dgs2tzx6p"
          },
          {
            "name": "order",
            "in": "query",
            "description": "The order in which the results are sorted. Supported values: asc, desc",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "string",
                  "default": "asc",
                  "enum": [
                    "asc",
                    "desc"
                  ]
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of results per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Return only transactions created on or after a specific height",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Return only transactions created on or before a specific height",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true,
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string, use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "inscription_id",
            "in": "query",
            "description": "Return only transactions containing a specific inscription, specified by an inscription ID. In presence of activity_kind, it relates to this specific inscription. In presence of exclude_self_transfers, it is this specific inscription that should be sent or received but not self-transferred.",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "example": "6fb976ab49dcec017f1e201e84395983204ae1a7c2abf7ced0a85d692e442799i0"
          },
          {
            "name": "activity_kind",
            "in": "query",
            "description": "Filter txs by presence of specific activity kind. Supported values: send, receive, self_transfer. In presence of inscription filter, the activity kind relates to that specific inscription. In presence of exclude_self_transfers, this activity kind cannot be self_transfer.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/InscriptionActivityKindByAddress"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "exclude_self_transfers",
            "in": "query",
            "description": "Exclude txs only containing inscriptions self-transfers. In presence of activity_kind, it cannot be self_transfer. In presence of inscription filter, that specific inscription should be sent or received, not self-transferred.",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedInscriptionActivityByAddress"
                },
                "example": {
                  "data": [
                    {
                      "confirmations": 60911,
                      "height": 836717,
                      "received": [
                        {
                          "from": {
                            "address": "bc1p4at29alvmmtunc5ffcpm9n5e4rvz63hlrjcrgd2zh7d7jy24xgns2g7rw7",
                            "input_index": 0,
                            "sat_offset": 0,
                            "script_pubkey": "5120af56a2f7ecded7c9e2894e03b2ce99a8d82d46ff1cb0343542bf9be911553227"
                          },
                          "inscription_id": "696936486c96124784d46dec66dd9e806280a182d1addf9763ff4bb92d0bc918i64",
                          "to": {
                            "address": "bc1p27j3fa2mr3d50m3uaavr0ntyzr0v2a27n48lc9gxpkzd4xye6dgs2tzx6p",
                            "output_txid": "a977a36055ba3d26203c82e6ee585539cadd12c3d17b236ffb67c051320b8991",
                            "output_vout": 0,
                            "sat_offset": 0,
                            "script_pubkey": "512057a514f55b1c5b47ee3cef5837cd6410dec5755e9d4ffc15060d84da9899d351"
                          }
                        }
                      ],
                      "self_transferred": [],
                      "sent": [],
                      "tx_hash": "a977a36055ba3d26203c82e6ee585539cadd12c3d17b236ffb67c051320b8991"
                    },
                    {
                      "confirmations": 60910,
                      "height": 836718,
                      "received": [],
                      "self_transferred": [],
                      "sent": [
                        {
                          "from": {
                            "address": "bc1pjuewj3dd4kpjen4zaqxd354jnwdjunewu7ncuaqqgsr0xa4cczzseg9066",
                            "input_index": 2,
                            "sat_offset": 0,
                            "script_pubkey": "51209732e945adad832ccea2e80cd8d2b29b9b2e4f2ee7a78e74004406f376b8c085"
                          },
                          "inscription_id": "696936486c96124784d46dec66dd9e806280a182d1addf9763ff4bb92d0bc918i64",
                          "to": {
                            "address": "bc1pjuewj3dd4kpjen4zaqxd354jnwdjunewu7ncuaqqgsr0xa4cczzseg9066",
                            "output_txid": "7647d53756d6f03b5191baaca26d84dbe2715406912eb2543c9d19c892a29c73",
                            "output_vout": 1,
                            "sat_offset": 0,
                            "script_pubkey": "51209732e945adad832ccea2e80cd8d2b29b9b2e4f2ee7a78e74004406f376b8c085"
                          }
                        }
                      ],
                      "tx_hash": "7647d53756d6f03b5191baaca26d84dbe2715406912eb2543c9d19c892a29c73"
                    }
                  ],
                  "last_updated": {
                    "block_hash": "0000000000000000000119bd8dffd7d8285a69744011aa98f0d9091b0555ca46",
                    "block_height": 897627
                  },
                  "next_cursor": null
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainTip": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string",
            "description": "The hash of the block",
            "example": "0000000000000000000a7f3b7b6b6e1d9a18db65a3b4a3f4f3bcb2e1f1b2d3e7"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "description": "The height of the block in the blockchain",
            "example": 707000,
            "minimum": 0
          }
        }
      },
...
      "FromInscriptionLocation": {
        "type": "object",
        "required": [
          "script_pubkey",
          "input_index",
          "sat_offset"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Address-encoding of the script pubkey at which the input containing the inscription resides.",
            "nullable": true
          },
          "input_index": {
            "type": "integer",
            "format": "int32",
            "description": "Index of the input containing the inscription.",
            "minimum": 0
          },
          "sat_offset": {
            "type": "integer",
            "format": "int64",
            "description": "Offset of the inscribed satoshi within the input.",
            "minimum": 0
          },
          "script_pubkey": {
            "type": "string",
            "description": "Script pubkey at which the input containing the inscription resides."
          }
        }
      },
...
      "InscriptionActivity": {
        "type": "object",
        "required": [
          "self_transferred",
          "sent",
          "received"
        ],
        "properties": {
          "received": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InscriptionActivityByTx"
            },
            "description": "List of inscriptions which the script gained control over."
          },
          "self_transferred": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InscriptionActivityByTx"
            },
            "description": "List of inscriptions that were self-transferred in the transaction."
          },
          "sent": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InscriptionActivityByTx"
            },
            "description": "List of inscriptions which the script lost control of."
          }
        }
      },
      "InscriptionActivityByAddress": {
        "type": "object",
        "required": [
          "height",
          "confirmations",
          "tx_hash",
          "inscription_activity"
        ],
        "properties": {
          "confirmations": {
            "type": "integer",
            "format": "int64",
            "description": "Number of confirmation blocks.",
            "minimum": 0
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "description": "Height of block containing the inscription activity.",
            "minimum": 0
          },
          "inscription_activity": {
            "$ref": "#/components/schemas/InscriptionActivity"
          },
          "tx_hash": {
            "type": "string",
            "description": "Hash of transaction containing the inscription activity."
          }
        }
      },
...
      "InscriptionActivityByTx": {
        "type": "object",
        "required": [
          "inscription_id",
          "to"
        ],
        "properties": {
          "from": {
            "allOf": [
              {
                "$ref": "#/components/schemas/FromInscriptionLocation"
              }
            ],
            "nullable": true
          },
          "inscription_id": {
            "type": "string",
            "description": "String representation of the inscription ID, whose first coordinate is the reveal transaction hash, and the second coordinate is the index of inscription in the reveal transaction."
          },
          "to": {
            "$ref": "#/components/schemas/ToInscriptionLocation"
          }
        }
      },
      "InscriptionActivityKindByAddress": {
        "type": "string",
        "enum": [
          "self_transfer",
          "send",
          "receive"
        ]
      },
...
      "PaginatedInscriptionActivityByAddress": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InscriptionActivityByAddress"
            }
          },
          "last_updated": {
            "$ref": "#/components/schemas/ChainTip"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ToInscriptionLocation": {
        "type": "object",
        "required": [
          "script_pubkey",
          "output_vout",
          "sat_offset",
          "output_txid"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "Address-encoding of the script pubkey at which the output containing the inscription resides.",
            "nullable": true
          },
          "output_txid": {
            "type": "string",
            "description": "Hash of tx producing the output containing the inscription. If this is the hash of the coinbase tx of the block, then the inscription was spent as fee in a tx in the block and therefore sent to the output of the coinbase tx controlled by the block miner."
          },
          "output_vout": {
            "type": "integer",
            "format": "int32",
            "description": "Index of the output containing the inscription.",
            "minimum": 0
          },
          "sat_offset": {
            "type": "integer",
            "format": "int64",
            "description": "Offset of the inscribed satoshi within the output.",
            "minimum": 0
          },
          "script_pubkey": {
            "type": "string",
            "description": "Script pubkey at which the output containing the inscription resides."
          }
        }
      },
...
    }
  }
}